Move tests to dataset_filters and update to remove mesh length fraction
